+++ .eslintignore
@@ -1,5 +1,3 @@
-lib/
-es/
 **/node_modules/**
 **/fixtures/**
 **/package.json
