+++ package.json
@@ -4,7 +4,6 @@
   "repository": "https://github.com/4Catalyzer/init.git",
   "author": "4Catalyzer",
   "license": "MIT",
-
   "private": true,
   "scripts": {
     "test": "npm run lint",
@@ -36,29 +35,17 @@
     "packages/*"
   ],
   "devDependencies": {
-    "@babel/cli": "^7.0.0-beta.39",
     "@babel/core": "^7.0.0-beta.39",
-    "babel-eslint": "^8.2.1",
+    "babel-eslint": "^9.0.0",
     "eslint": "^4.16.0",
-    "eslint-config-4catalyzer": "^0.3.3",
-    "eslint-config-prettier": "^2.9.0",
+    "eslint-config-4catalyzer": "^0.4.4",
+    "eslint-config-prettier": "^3.0.1",
     "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-jest": "^21.7.0",
     "eslint-plugin-prettier": "^2.5.0",
     "husky": "^0.14.3",
     "lerna": "^3.3.0",
-    "lint-staged": "^6.1.0",
+    "lint-staged": "^7.2.2",
     "prettier": "^1.10.2"
-  },
-  "dependencies": {
-    "@4c/file-butler": "^1.2.0",
-    "chalk": "^2.3.0",
-    "execa": "^0.9.0",
-    "fs-extra": "^5.0.0",
-    "inquirer": "^5.0.1",
-    "ora": "^1.4.0",
-    "semver": "^5.5.0",
-    "slash": "^1.0.0",
-    "yargs": "^11.0.0"
   }
 }
