--- packages/4c-file-butler/.eslintignore
@@ -1,6 +1,0 @@
-lib/
-es/
-**/node_modules/**
-**/fixtures/**
-**/package.json
-**/CHANGELOG.md
