--- packages/4c-file-butler/.eslintrc
@@ -1,30 +1,0 @@
-{
-  "extends": ["4catalyzer", "prettier"],
-  "plugins": ["prettier"],
-  "env": {
-    "node": true
-  },
-  "rules": {
-    "prettier/prettier": "error",
-    "global-require": "off",
-    "import/no-dynamic-require": "off"
-  },
-  "overrides": [
-    {
-      "files": ["test/**"],
-      "plugins": ["jest"],
-      "env": {
-        "jest/globals": true
-      },
-      "rules": {
-        "no-console": "off",
-        "no-await-in-loop": "off",
-        "jest/no-disabled-tests": "warn",
-        "jest/no-focused-tests": "error",
-        "jest/no-identical-title": "error",
-        "jest/prefer-to-have-length": "warn",
-        "jest/valid-expect": "error"
-      }
-    }
-  ]
-}
