--- packages/4c-file-butler/.travis.yml
@@ -1,12 +1,0 @@
-sudo: false
-
-language: node_js
-node_js:
-  - '10'
-  - '8'
-cache: yarn
-after_success:
-  - yarn travis-deploy-once "yarn semantic-release"
-branches:
-  only:
-    - master
