--- packages/4c-file-butler/package.json
@@ -1,72 +1,0 @@
-{
-  "name": "@4c/file-butler",
-  "version": "1.1.1",
-  "main": "lib.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/4Catalyzer/file-butler.git"
-  },
-  "author": "4Catalyzer",
-  "license": "MIT",
-  "scripts": {
-    "tdd": "jest --watch",
-    "test": "npm run lint",
-    "testonly": "jest",
-    "lint":
-      "eslint . && prettier --list-different --ignore-path .eslintignore '**/*.{json,css,md}'",
-    "format":
-      "eslint . --fix && prettier --write --ignore-path .eslintignore '**/*.{json,css,md}'",
-    "precommit": "lint-staged"
-  },
-  "bin": {
-    "file-butler": "index.js"
-  },
-  "files": ["lib.js", "index.js"],
-  "engines": {
-    "node": ">= 8.3"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "prettier": {
-    "printWidth": 79,
-    "singleQuote": true,
-    "trailingComma": "all"
-  },
-  "lint-staged": {
-    "*.js": ["eslint --fix", "git add"],
-    "*.{json,css,md}": [
-      "prettier --write --ignore-path .eslintignore",
-      "git add"
-    ]
-  },
-  "jest": {
-    "roots": ["<rootDir>/test"],
-    "testEnvironment": "node"
-  },
-  "release": {
-    "extends": ["@4c/semantic-release-config"]
-  },
-  "devDependencies": {
-    "@4c/semantic-release-config": "^1.0.2",
-    "babel-eslint": "^8.2.1",
-    "eslint": "^4.16.0",
-    "eslint-config-4catalyzer": "^0.4.1",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jest": "^21.7.0",
-    "eslint-plugin-prettier": "^2.5.0",
-    "husky": "^0.14.3",
-    "jest": "^22.4.4",
-    "lint-staged": "^7.1.0",
-    "prettier": "^1.10.2",
-    "semantic-release": "^15.5.0",
-    "travis-deploy-once": "^5.0.0"
-  },
-  "dependencies": {
-    "cpy": "^7.0.0",
-    "fs-extra": "^7.0.0",
-    "globby": "^8.0.1",
-    "yargs": "^11.0.0"
-  }
-}
