+++ packages/cli/index.js
@@ -1,8 +1,14 @@
 #!/usr/bin/env node
 
 const yargs = require('yargs');
-const newCommand = require('./commands/new');
-const releaseCommand = require('./commands/release');
+
+const init = require('@4c/init/command');
+const rollout = require('@4c/rollout/command');
+
+const setCmdName = (name, cmd) => ({
+  ...cmd,
+  command: cmd.command.replace(/^\$0/, name),
+});
 
 yargs
   .help()
@@ -13,6 +19,6 @@ yargs
   .demandCommand(1, `Pass --help to see all available commands and options.`)
   .strict()
   .recommendCommands()
-  .command(newCommand)
-  .command(releaseCommand)
+  .command(setCmdName('init', init))
+  .command(setCmdName('release', rollout))
   .parse(process.argv.slice(2));
