+++ packages/cli/package.json
@@ -8,57 +8,18 @@
   "bin": {
     "4c": "index.js"
   },
-  "scripts": {
-    "test": "npm run lint",
-    "lint": "eslint . && prettier --list-different --ignore-path .eslintignore '**/*.{json,css,md}'",
-    "precommit": "lint-staged"
-  },
-  "prettier": {
-    "printWidth": 79,
-    "singleQuote": true,
-    "trailingComma": "all"
-  },
-  "importSort": {
-    ".js": {
-      "parser": "4catalyzer",
-      "style": "4catalyzer"
-    }
-  },
-  "lint-staged": {
-    "*.js": [
-      "eslint --fix",
-      "git add"
-    ],
-    "*.{json,css,md}": [
-      "prettier --write --ignore-path .eslintignore",
-      "git add"
-    ]
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.0.0-beta.39",
-    "@babel/core": "^7.0.0-beta.39",
-    "babel-eslint": "^8.2.1",
-    "eslint": "^4.16.0",
-    "eslint-config-4catalyzer": "^0.3.3",
-    "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jest": "^21.7.0",
-    "eslint-plugin-prettier": "^2.5.0",
-    "husky": "^0.14.3",
-    "lerna": "^3.3.0",
-    "lint-staged": "^6.1.0",
-    "prettier": "^1.10.2"
-  },
   "dependencies": {
-    "@4c/file-butler": "^1.2.0",
     "@4c/cli-core": "^1.0.0",
+    "@4c/file-butler": "^1.2.0",
+    "@4c/init": "^1.0.0",
+    "@4c/rollout": "^1.0.0",
     "chalk": "^2.3.0",
-    "execa": "^0.9.0",
-    "fs-extra": "^5.0.0",
-    "inquirer": "^5.0.1",
-    "ora": "^1.4.0",
+    "execa": "^1.0.0",
+    "fs-extra": "^7.0.0",
+    "inquirer": "^6.2.0",
+    "ora": "^3.0.0",
     "semver": "^5.5.0",
-    "slash": "^1.0.0",
-    "yargs": "^11.0.0"
+    "slash": "^2.0.0",
+    "yargs": "^12.0.1"
   }
 }
