+++ packages/init/package.json
@@ -5,13 +5,13 @@
   "author": "4catalyzer",
   "license": "MIT",
   "bin": {
-    "4c": "cli.js"
+    "init": "cli.js"
   },
   "dependencies": {
     "@4c/cli-core": "^1.0.0",
-    "execa": "^0.9.0",
-    "fs-extra": "^5.0.0",
-    "inquirer": "^5.0.1",
+    "execa": "^1.0.0",
+    "fs-extra": "^7.0.0",
+    "inquirer": "^6.2.0",
     "yargs": "^12.0.1"
   }
 }
