+++ packages/rollout/command.js
@@ -1,242 +1,291 @@
 const path = require('path');
 const fs = require('fs-extra');
+const chalk = require('chalk');
 const execa = require('execa');
-const inquirer = require('inquirer');
+const semver = require('semver');
+const { createAltPublishDir } = require('@4c/file-butler');
 
 const GitUtilities = require('@4c/cli-core/GitUtilities');
+const ConsoleUtilities = require('@4c/cli-core/ConsoleUtilities');
+const PromptUtilities = require('@4c/cli-core/PromptUtilities');
 
-const templatePath = path.resolve(__dirname, './templates');
+const writeJson = (p, json) => fs.writeJson(p, json, { spaces: 2 });
 
-async function setupNpm(dest, a) {
-  const eslint = {
-    eslint: '^4.16.0',
-    'eslint-config-prettier': '^2.9.0',
-    'eslint-plugin-import': '^2.8.0',
-    'eslint-plugin-prettier': '^2.5.0',
-    'eslint-plugin-jest': '^21.7.0',
-  };
+async function runLifecycle(script, pkg) {
+  if (!pkg.scripts || !pkg.scripts[script]) return;
+  await execa('npm', ['run', script], { stdio: [0, 1, 'pipe'] });
+}
 
-  if (a.type === 'web') {
-    eslint['eslint-config-4catalyzer-react'] = '^0.4.1';
-    eslint['eslint-plugin-react'] = '^7.5.1';
-    eslint['eslint-plugin-jsx-a11y'] = '^6.0.3';
-  } else {
-    eslint['eslint-config-4catalyzer'] = '^0.4.1';
-  }
-
-  fs.writeJSON(
-    path.join(dest, 'package.json'),
-    {
-      name: a.name,
-      version: '1.0.0',
-      main: a.babel ? 'lib/index.js' : 'index.js',
-      ...(a.babel && {
-        module: 'es/index.js',
-      }),
-
-      repository: {
-        type: 'git',
-        url: GitUtilities.remoteUrl(a.name),
-      },
-      author: '4Catalyzer',
-      license: 'MIT',
-      scripts: {
-        tdd: 'jest --watch',
-        test: 'npm run lint && jest',
-        testonly: 'jest',
-        ...(a.babel && {
-          'build:es':
-            'babel src -d es --env-name esm --ignore **/__tests__ --delete-dir-on-start',
-          'build:lib':
-            'babel src -d lib --ignore **/__tests__ --delete-dir-on-start',
-          build: 'npm run build:lib && npm run build:es',
-          prepublishOnly: 'yarn run build',
-        }),
-        lint: [
-          'eslint .',
-          "prettier --list-different --ignore-path .eslintignore '**/*.{json,css,md}'",
-        ].join(' && '),
-        format: [
-          'eslint . --fix',
-          "prettier --write --ignore-path .eslintignore '**/*.{json,css,md}'",
-        ].join(' && '),
-        precommit: 'lint-staged',
-      },
-      publishConfig: {
-        access: a.isPrivate ? 'restricted' : 'public',
-      },
-      prettier: {
-        printWidth: 79,
-        singleQuote: true,
-        trailingComma: 'all',
-      },
-      'lint-staged': {
-        '*.js': ['eslint --fix', 'git add'],
-        '*.{json,css,md}': [
-          'prettier --write --ignore-path .eslintignore',
-          'git add',
-        ],
-      },
-      jest: {
-        roots: ['<rootDir>/test'],
-        testEnvironment: a.type === 'node' ? 'node' : 'jsdom',
-      },
-      ...(a.semanticRelease && {
-        release: {
-          extends: ['@4c/semantic-release-config'],
-        },
-      }),
-      devDependencies: {
-        ...(a.babel && {
-          '@babel/cli': '^7.0.0-beta.39',
-          '@babel/core': '^7.0.0-beta.39',
-          '@4c/babel-preset-4catalyzer': '^1.0.0',
-          'babel-jest': '^22.4.3',
-          'babel-core': 'bridge',
-        }),
-        ...(a.semanticRelease && {
-          '@4c/semantic-release-config': '^1.0.2',
-        }),
-        'babel-eslint': '^8.2.1',
-        husky: '^0.14.3',
-        'lint-staged': '^7.1.0',
-        prettier: '^1.10.2',
-        jest: '^22.4.4',
-        ...eslint,
-      },
+async function maybeRollbackGit(tag, skipGit, skipVersion) {
+  if (skipGit) return;
+  const confirmed = await PromptUtilities.confirm(
+    'There was a problem publishing, do you want to rollback the git operations?',
+  );
+  await ConsoleUtilities.step(
+    `Rolling back git operations`,
+    async () => {
+      await GitUtilities.removeTag(tag);
+      if (!skipVersion) await GitUtilities.removeLastCommit();
     },
-    { spaces: 2 },
+    !confirmed,
   );
 }
 
-exports.command = 'new [location]';
+async function bumpVersion(pkg, publishDir, cwd) {
+  const json = pkg;
+  const pkgPath = path.join(cwd, 'package.json');
+  await writeJson(pkgPath, json);
 
-exports.describe = 'create a new package';
+  if (publishDir) {
+    await createAltPublishDir({ outDir: publishDir });
+  }
+}
+
+async function getNextVersion(version, currentVersion, preid) {
+  const patch = semver.inc(currentVersion, 'patch');
+  const minor = semver.inc(currentVersion, 'minor');
+  const major = semver.inc(currentVersion, 'major');
+  const prepatch = semver.inc(currentVersion, 'prepatch');
+  const preminor = semver.inc(currentVersion, 'preminor');
+  const premajor = semver.inc(currentVersion, 'premajor');
+
+  if (semver.valid(version)) {
+    return version;
+  }
+  switch (version) {
+    case 'patch':
+      return patch;
+    case 'minor':
+      return minor;
+    case 'major':
+      return major;
+    default:
+  }
+
+  const message = `Select a new version (currently ${currentVersion})`;
+
+  const choice = await PromptUtilities.select(message, {
+    choices: [
+      { value: patch, name: `Patch (${patch})` },
+      { value: minor, name: `Minor (${minor})` },
+      { value: major, name: `Major (${major})` },
+      { value: prepatch, name: `Prepatch (${prepatch})` },
+      { value: preminor, name: `Preminor (${preminor})` },
+      { value: premajor, name: `Premajor (${premajor})` },
+      { value: 'PRERELEASE', name: 'Prerelease' },
+      { value: 'CUSTOM', name: 'Custom' },
+    ],
+  });
+
+  switch (choice) {
+    case 'CUSTOM': {
+      return PromptUtilities.input('Enter a custom version', {
+        filter: semver.valid,
+        validate: v => v !== null || 'Must be a valid semver version',
+      });
+    }
+
+    case 'PRERELEASE': {
+      const [existingId] = semver.prerelease(currentVersion) || [];
+      const defaultVersion = semver.inc(
+        currentVersion,
+        'prerelease',
+        existingId,
+      );
+      const prompt = `(default: ${
+        existingId ? `"${existingId}"` : 'none'
+      }, yielding ${defaultVersion})`;
+
+      const nextPreId =
+        preid !== 'latest'
+          ? preid
+          : await PromptUtilities.input(
+              `Enter a prerelease identifier ${prompt}`,
+            );
+
+      return semver.inc(currentVersion, 'prerelease', nextPreId);
+    }
+
+    default: {
+      return choice;
+    }
+  }
+}
+
+exports.command = '$0 [nextVersion]';
+
+exports.describe = 'Publish a new version';
 
 exports.builder = _ =>
-  _.positional('location', {
+  _.positional('nextVersion', {
     type: 'string',
-    default: process.cwd(),
-    describe: 'the location of the package',
-    normalize: true,
-  });
-
-exports.handler = async ({ location, cwd = process.cwd() }) => {
-  const dest = path.isAbsolute(location)
-    ? location
-    : path.resolve(cwd, location);
-
-  const copyTemplate = (src, destName = src) =>
-    fs.copyFile(path.join(templatePath, src), path.join(dest, destName));
-
-  // my own convention of naming scoped repo's like 4c-foo on disk
-  const getName = ({ scope }) => {
-    let name = path.basename(dest);
-
-    if (!scope) return name;
-    name = name.replace(new RegExp(`^${scope.slice(1)}-`), '');
-    return `${scope}/${name}`;
-  };
-
-  const answers = await inquirer.prompt([
-    {
-      name: 'scopePackage',
-      type: 'confirm',
-      message: 'Create a scoped package?',
-      default: true,
-    },
-    {
-      name: 'scope',
-      type: 'input',
-      message: 'package scope',
-      default: '@4c',
-      when: _ => !!_.scopePackage,
-    },
-    {
-      name: 'isPrivate',
-      type: 'confirm',
+    describe: 'The next version',
+  })
+    .pkgConf('release')
+    .option('preid', {
+      type: 'string',
+    })
+    .option('prerelease', {
+      type: 'bool',
       default: false,
-      message: 'Is this a private package?',
-      when: _ => !!_.scope,
-    },
-    {
-      name: 'name',
-      type: 'input',
-      message: 'name',
-      default: getName,
-    },
-    {
-      name: 'type',
-      type: 'list',
-      default: 'node',
-      choices: ['node', 'web'],
-      message: 'What type of library is this?',
-    },
-    {
-      name: 'babel',
-      type: 'confirm',
-      default: false,
-      message: 'Do you need babel (maybe not?)',
-      when: _ => _.type === 'node',
-    },
-    {
-      name: 'semanticRelease',
-      type: 'confirm',
-      default: true,
-      message: 'Do you want to use semantic-release to handle releases?',
-    },
-  ]);
+    })
+    .option('publish-dir', {
+      type: 'string',
+    })
+    .option('allow-branch', {
+      group: 'Command Options:',
+      describe: 'Specify which branches to allow publishing from.',
+      type: 'array',
+      default: ['master'],
+    })
+    .option('npm-tag', {
+      type: 'string',
+    })
+    .option('skip-version', {
+      describe: 'Skip version bumping',
+      type: 'boolean',
+    })
+    .option('skip-checks', {
+      describe: 'Skip tests, linting and git hygiene checks',
+      type: 'boolean',
+    })
+    .option('skip-git', {
+      describe: 'Skip commiting, tagging, and pushing git changes.',
+      type: 'boolean',
+    })
+    .option('skip-npm', {
+      describe: 'Stop before actually publishing change to npm.',
+      type: 'boolean',
+    })
+    .option('public', {
+      type: 'bool',
+      default: undefined,
+    });
 
-  if (answers.type === 'web') {
-    answers.babel = true;
-  }
+exports.handler = async ({
+  preid,
+  nextVersion: version,
+  npmTag,
+  skipChecks,
+  skipGit,
+  skipNpm,
+  skipVersion,
+  allowBranch,
+  publishDir,
+  public: isPublic,
+}) => {
+  const cwd = process.cwd();
+  const pkgPath = path.join(cwd, 'package.json');
+  const pkg = require(pkgPath);
 
-  await fs.ensureDir(dest);
+  try {
+    await ConsoleUtilities.step(
+      'Checking repo and running tests',
+      async () => {
+        if (!skipGit) {
+          await GitUtilities.assertClean();
+          await GitUtilities.assertMatchesRemote();
+        }
+        const branch = await GitUtilities.getCurrentBranch();
 
-  await copyTemplate('gitignore', '.gitignore');
-  await copyTemplate('.travis.yml');
-  await copyTemplate(`${answers.type}.eslintrc`, '.eslintrc');
-  await copyTemplate('.eslintignore');
-  await copyTemplate('LICENSE');
+        if (!allowBranch.includes(branch))
+          throw new Error(`Cannot publish from branch: ${chalk.bold(branch)}`);
 
-  await GitUtilities.init(dest);
-  await GitUtilities.addRemote(dest, answers.name);
-
-  await setupNpm(dest, answers);
-
-  if (answers.babel) {
-    await fs.ensureFile(path.join(dest, 'src/index.js'));
-    await fs.writeFile(
-      path.join(dest, '.babelrc.js'),
-      `
-module.exports = api => ({
-  presets: [
-    [
-      '@4c/4catalyzer',
-      {
-        target: '${answers.type}',
-        modules: api.env() === 'esm' ? false : 'commonjs'
+        await execa('npm', ['test'], { stdio: [0, 1, 'pipe'] });
       },
-    ],
-  ]
-});
-      `,
+      skipChecks,
     );
-  } else {
-    await fs.ensureFile(path.join(dest, 'index.js'));
-  }
 
-  await execa('yarn', ['install'], { cwd: dest, stdio: 'inherit' });
-  await execa('yarn', ['upgrade-interactive', '--latest'], {
-    cwd: dest,
-    stdio: 'inherit',
-  });
+    let nextVersion = pkg.version;
 
-  if (answers.semanticRelease) {
+    if (skipVersion) {
+      ConsoleUtilities.spinner().warn(
+        `Using existing version: ${chalk.bold(nextVersion)}`,
+      );
+    } else {
+      nextVersion = await getNextVersion(version, pkg.version, preid);
+
+      await ConsoleUtilities.step(
+        `Bumping version to: ${chalk.bold(nextVersion)}  (${chalk.dim(
+          `was ${pkg.version}`,
+        )})`,
+        () => bumpVersion({ ...pkg, version: nextVersion }, publishDir, cwd),
+      );
+    }
+
+    const isPrerelease = !!semver.prerelease(nextVersion);
+    const tag = npmTag || isPrerelease ? 'next' : 'latest';
+
+    const confirmed = await PromptUtilities.confirm(
+      `Are you sure you want to publish version: ${nextVersion}@${tag}?`,
+    );
+
+    if (!confirmed) return;
+    const gitTag = `v${nextVersion}`;
+
+    await ConsoleUtilities.step(
+      'Tagging and committing version bump',
+      async () => {
+        if (!skipVersion) {
+          await GitUtilities.addFile(pkgPath);
+          await GitUtilities.commit(`Publish ${gitTag}`);
+        }
+        await GitUtilities.addTag(gitTag);
+      },
+      skipGit,
+    );
+
+    try {
+      await ConsoleUtilities.step(
+        'Publishing to npm',
+        async () => {
+          const args = ['publish'];
+          if (publishDir) {
+            args.push(publishDir);
+            // We run the lifecycle scripts manually to ensure they run in
+            // the package root, not the publish dir
+            await runLifecycle('prepublish', pkg);
+            await runLifecycle('prepare', pkg);
+            await runLifecycle('prepublishOnly', pkg);
+          }
+          if (tag !== 'latest') {
+            args.push('--tag', tag);
+          }
+
+          if (isPublic != null) {
+            args.push('--access', isPublic ? 'public' : 'restricted');
+          }
+          await execa('npm', args, { stdio: [0, 1, 'pipe'] });
+
+          try {
+            if (publishDir) {
+              await runLifecycle('publish', pkg);
+              await runLifecycle('postpublish', pkg);
+            }
+          } catch (err) {
+            console.error(err);
+            /* we've already published so we shouldn't try and rollback if these fail */
+          }
+        },
+        skipNpm,
+      );
+    } catch (err) {
+      await maybeRollbackGit(gitTag, skipGit, skipVersion);
+      throw err;
+    }
+
+    if (!skipGit) {
+      await GitUtilities.pushWithTags();
+    }
+
     console.log(
-      '\nRun `npx semantic-release-cli setup` after pushing to github for the first time to setup semantic release\n',
+      `🎉  Published v${nextVersion}@${tag}:  ${chalk.blue(
+        skipNpm
+          ? await GitUtilities.getRemoteUrl()
+          : `https://npm.im/${pkg.name}`,
+      )} \n`,
     );
-  } else {
-    console.log('Done!');
+  } catch (err) {
+    /* ignore */
   }
 };
